requirements-cli.txt
# CLI 工具依賴
click>=8.0.0
tabulate>=0.9.0
PyYAML>=6.0
colorama>=0.4.4

# 開發和測試依賴
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0 